HW4 done
