+++ service/settings.py
@@ -25,10 +25,11 @@ class ServiceConfig(Config):
     service_name: str = "reco_service"
     k_recs: int = 10
     api_key: str
-    zip_knn_path: str = 'models/knn.zip'
+    zip_models_path: str = 'models/models.zip'
     knn_model: str = 'knn_bm25_item.dill'
     pop_model: str = 'pop_model_7.dill'
     users_list: str = 'users_list.pickle'
+    emb_maps: str = 'emb_maps.pickle'
 
     log_config: LogConfig
